Refactored Bucket
