+++ src/bucket/BucketApplicator.h
@@ -4,8 +4,8 @@
 // under the Apache License, Version 2.0. See the COPYING file at the root
 // of this distribution or at http://www.apache.org/licenses/LICENSE-2.0
 
-#include "bucket/Bucket.h"
 #include "bucket/BucketInputIterator.h"
+#include "bucket/LiveBucket.h"
 #include "util/Timer.h"
 #include <memory>
