+++ src/bucket/BucketIndexImpl.h
@@ -4,8 +4,8 @@
 // under the Apache License, Version 2.0. See the COPYING file at the root
 // of this distribution or at http://www.apache.org/licenses/LICENSE-2.0
 
-#include "bucket/Bucket.h"
 #include "bucket/BucketIndex.h"
+#include "bucket/LiveBucket.h"
 #include "medida/meter.h"
 #include "util/BinaryFuseFilter.h"
 #include "xdr/Stellar-types.h"
