+++ src/bucket/BucketInputIterator.cpp
@@ -3,7 +3,8 @@
 // of this distribution or at http://www.apache.org/licenses/LICENSE-2.0
 
 #include "bucket/BucketInputIterator.h"
-#include "bucket/Bucket.h"
+#include "bucket/HotArchiveBucket.h"
+#include "bucket/LiveBucket.h"
 #include "xdr/Stellar-ledger.h"
 #include <Tracy.hpp>
 #include <type_traits>
