+++ src/bucket/FutureBucket.h
@@ -4,7 +4,6 @@
 // under the Apache License, Version 2.0. See the COPYING file at the root
 // of this distribution or at http://www.apache.org/licenses/LICENSE-2.0
 
-#include "bucket/Bucket.h"
 #include "bucket/BucketUtils.h"
 #include "util/GlobalChecks.h"
 #include <cereal/cereal.hpp>
