+++ src/bucket/MergeKey.h
@@ -3,7 +3,6 @@
 // Copyright 2019 Stellar Development Foundation and contributors. Licensed
 // under the Apache License, Version 2.0. See the COPYING file at the root
 // of this distribution or at http://www.apache.org/licenses/LICENSE-2.0
-#include "bucket/Bucket.h"
 #include "xdr/Stellar-types.h"
 #include <iosfwd>
 #include <vector>
