+++ src/bucket/test/BucketTests.cpp
@@ -11,10 +11,11 @@
 // first to include <windows.h> -- so we try to include it before everything
 // else.
 #include "util/asio.h"
-#include "bucket/Bucket.h"
 #include "bucket/BucketInputIterator.h"
 #include "bucket/BucketManager.h"
 #include "bucket/BucketOutputIterator.h"
+#include "bucket/HotArchiveBucket.h"
+#include "bucket/LiveBucket.h"
 #include "bucket/test/BucketTestUtils.h"
 #include "ledger/LedgerTxn.h"
 #include "ledger/test/LedgerTestUtils.h"
@@ -80,7 +81,7 @@ TEST_CASE_VERSIONS("file backed buckets", "[bucket][bucketbench]")
             dead = LedgerTestUtils::generateValidLedgerEntryKeysWithExclusions(
                 {CONFIG_SETTING}, 1000);
             {
-                b1 = Bucket::merge(
+                b1 = BucketBase::merge(
                     app->getBucketManager(),
                     app->getConfig().LEDGER_PROTOCOL_VERSION, b1,
                     LiveBucket::fresh(app->getBucketManager(),
@@ -169,11 +170,11 @@ TEST_CASE_VERSIONS("merging bucket entries", "[bucket]")
                                                /*countMergeEvents=*/true,
                                                clock.getIOContext(),
                                                /*doFsync=*/true);
-                auto b1 = Bucket::merge(bm, vers, bLive, bDead, /*shadows=*/{},
-                                        /*keepTombstoneEntries=*/true,
-                                        /*countMergeEvents=*/true,
-                                        clock.getIOContext(),
-                                        /*doFsync=*/true);
+                auto b1 = BucketBase::merge(
+                    bm, vers, bLive, bDead, /*shadows=*/{},
+                    /*keepTombstoneEntries=*/true,
+                    /*countMergeEvents=*/true, clock.getIOContext(),
+                    /*doFsync=*/true);
                 CHECK(countEntries(b1) == 1);
             }
         };
@@ -208,11 +209,11 @@ TEST_CASE_VERSIONS("merging bucket entries", "[bucket]")
                                            /*countMergeEvents=*/true,
                                            clock.getIOContext(),
                                            /*doFsync=*/true);
-            auto b1 =
-                Bucket::merge(bm, vers, bLive, bDead, /*shadows=*/{},
-                              /*keepTombstoneEntries=*/true,
-                              /*countMergeEvents=*/true, clock.getIOContext(),
-                              /*doFsync=*/true);
+            auto b1 = BucketBase::merge(bm, vers, bLive, bDead, /*shadows=*/{},
+                                        /*keepTombstoneEntries=*/true,
+                                        /*countMergeEvents=*/true,
+                                        clock.getIOContext(),
+                                        /*doFsync=*/true);
             EntryCounts e(b1);
             CHECK(e.sum() == live.size());
             CLOG_DEBUG(Bucket, "post-merge live count: {} of {}", e.nLive,
@@ -262,12 +263,12 @@ TEST_CASE_VERSIONS("merging bucket entries", "[bucket]")
                 app->getBucketManager(), getAppLedgerVersion(app), {}, live,
                 dead, /*countMergeEvents=*/true, clock.getIOContext(),
                 /*doFsync=*/true);
-            std::shared_ptr<LiveBucket> b3 =
-                Bucket::merge(app->getBucketManager(),
-                              app->getConfig().LEDGER_PROTOCOL_VERSION, b1, b2,
-                              /*shadows=*/{}, /*keepTombstoneEntries=*/true,
-                              /*countMergeEvents=*/true, clock.getIOContext(),
-                              /*doFsync=*/true);
+            std::shared_ptr<LiveBucket> b3 = BucketBase::merge(
+                app->getBucketManager(),
+                app->getConfig().LEDGER_PROTOCOL_VERSION, b1, b2,
+                /*shadows=*/{}, /*keepTombstoneEntries=*/true,
+                /*countMergeEvents=*/true, clock.getIOContext(),
+                /*doFsync=*/true);
             CHECK(countEntries(b3) == liveCount);
         }
     });
@@ -319,11 +320,11 @@ TEST_CASE_VERSIONS("merging hot archive bucket entries", "[bucket][archival]")
             // e2 -> ARCHIVED
             // e3 -> LIVE
             // e4 -> DELETED
-            auto merged =
-                Bucket::merge(bm, vers, b1, b2, /*shadows=*/{},
-                              /*keepTombstoneEntries=*/true,
-                              /*countMergeEvents=*/true, clock.getIOContext(),
-                              /*doFsync=*/true);
+            auto merged = BucketBase::merge(bm, vers, b1, b2, /*shadows=*/{},
+                                            /*keepTombstoneEntries=*/true,
+                                            /*countMergeEvents=*/true,
+                                            clock.getIOContext(),
+                                            /*doFsync=*/true);
 
             bool seen1 = false;
             bool seen4 = false;
@@ -462,11 +463,11 @@ TEST_CASE("merges proceed old-style despite newer shadows",
     {
         // With proto 12, new bucket version solely depends on the snap version
         auto bucket =
-            Bucket::merge(bm, v12, b11first, b11second,
-                          /*shadows=*/{b12first},
-                          /*keepTombstoneEntries=*/true,
-                          /*countMergeEvents=*/true, clock.getIOContext(),
-                          /*doFsync=*/true);
+            BucketBase::merge(bm, v12, b11first, b11second,
+                              /*shadows=*/{b12first},
+                              /*keepTombstoneEntries=*/true,
+                              /*countMergeEvents=*/true, clock.getIOContext(),
+                              /*doFsync=*/true);
         REQUIRE(bucket->getBucketVersion() == v11);
     }
     SECTION("shadow versions mixed, pick lower")
@@ -474,21 +475,21 @@ TEST_CASE("merges proceed old-style despite newer shadows",
         // Merging older version (10) buckets, with mixed versions of shadows
         // (11, 12) Pick initentry (11) style merge
         auto bucket =
-            Bucket::merge(bm, v12, b10first, b10second,
-                          /*shadows=*/{b12first, b11second},
-                          /*keepTombstoneEntries=*/true,
-                          /*countMergeEvents=*/true, clock.getIOContext(),
-                          /*doFsync=*/true);
+            BucketBase::merge(bm, v12, b10first, b10second,
+                              /*shadows=*/{b12first, b11second},
+                              /*keepTombstoneEntries=*/true,
+                              /*countMergeEvents=*/true, clock.getIOContext(),
+                              /*doFsync=*/true);
         REQUIRE(bucket->getBucketVersion() == v11);
     }
     SECTION("refuse to merge new version with shadow")
     {
-        REQUIRE_THROWS_AS(Bucket::merge(bm, v12, b12first, b12second,
-                                        /*shadows=*/{b12first},
-                                        /*keepTombstoneEntries=*/true,
-                                        /*countMergeEvents=*/true,
-                                        clock.getIOContext(),
-                                        /*doFsync=*/true),
+        REQUIRE_THROWS_AS(BucketBase::merge(bm, v12, b12first, b12second,
+                                            /*shadows=*/{b12first},
+                                            /*keepTombstoneEntries=*/true,
+                                            /*countMergeEvents=*/true,
+                                            clock.getIOContext(),
+                                            /*doFsync=*/true),
                           std::runtime_error);
     }
 }
@@ -519,12 +520,12 @@ TEST_CASE("merges refuse to exceed max protocol version",
         LiveBucket::fresh(bm, vers, {}, {otherLiveA}, {},
                           /*countMergeEvents=*/true, clock.getIOContext(),
                           /*doFsync=*/true);
-    REQUIRE_THROWS_AS(Bucket::merge(bm, vers - 1, bnew1, bnew2,
-                                    /*shadows=*/{},
-                                    /*keepTombstoneEntries=*/true,
-                                    /*countMergeEvents=*/true,
-                                    clock.getIOContext(),
-                                    /*doFsync=*/true),
+    REQUIRE_THROWS_AS(BucketBase::merge(bm, vers - 1, bnew1, bnew2,
+                                        /*shadows=*/{},
+                                        /*keepTombstoneEntries=*/true,
+                                        /*countMergeEvents=*/true,
+                                        clock.getIOContext(),
+                                        /*doFsync=*/true),
                       std::runtime_error);
 }
 
@@ -595,7 +596,7 @@ TEST_CASE_VERSIONS("merging bucket entries with initentry",
                                            /*countMergeEvents=*/true,
                                            clock.getIOContext(),
                                            /*doFsync=*/true);
-            auto b1 = Bucket::merge(
+            auto b1 = BucketBase::merge(
                 bm, cfg.LEDGER_PROTOCOL_VERSION, bInit, bDead, /*shadows=*/{},
                 /*keepTombstoneEntries=*/true,
                 /*countMergeEvents=*/true, clock.getIOContext(),
@@ -635,12 +636,12 @@ TEST_CASE_VERSIONS("merging bucket entries with initentry",
                                            /*countMergeEvents=*/true,
                                            clock.getIOContext(),
                                            /*doFsync=*/true);
-            auto bmerge1 = Bucket::merge(
+            auto bmerge1 = BucketBase::merge(
                 bm, cfg.LEDGER_PROTOCOL_VERSION, bInit, bLive, /*shadows=*/{},
                 /*keepTombstoneEntries=*/true,
                 /*countMergeEvents=*/true, clock.getIOContext(),
                 /*doFsync=*/true);
-            auto b1 = Bucket::merge(
+            auto b1 = BucketBase::merge(
                 bm, cfg.LEDGER_PROTOCOL_VERSION, bmerge1, bDead, /*shadows=*/{},
                 /*keepTombstoneEntries=*/true,
                 /*countMergeEvents=*/true, clock.getIOContext(),
@@ -705,12 +706,12 @@ TEST_CASE_VERSIONS("merging bucket entries with initentry",
             CHECK(enew.nLive == 0);
             CHECK(enew.nDead == 1);
 
-            auto bmerge1 = Bucket::merge(
+            auto bmerge1 = BucketBase::merge(
                 bm, cfg.LEDGER_PROTOCOL_VERSION, bold, bmed, /*shadows=*/{},
                 /*keepTombstoneEntries=*/true,
                 /*countMergeEvents=*/true, clock.getIOContext(),
                 /*doFsync=*/true);
-            auto bmerge2 = Bucket::merge(
+            auto bmerge2 = BucketBase::merge(
                 bm, cfg.LEDGER_PROTOCOL_VERSION, bmerge1, bnew, /*shadows=*/{},
                 /*keepTombstoneEntries=*/true,
                 /*countMergeEvents=*/true, clock.getIOContext(),
@@ -792,12 +793,12 @@ TEST_CASE_VERSIONS("merging bucket entries with initentry with shadows",
                                         /*countMergeEvents=*/true,
                                         clock.getIOContext(),
                                         /*doFsync=*/true);
-            auto merged =
-                Bucket::merge(bm, cfg.LEDGER_PROTOCOL_VERSION, b1, b2,
-                              /*shadows=*/{shadow},
-                              /*keepTombstoneEntries=*/true,
-                              /*countMergeEvents=*/true, clock.getIOContext(),
-                              /*doFsync=*/true);
+            auto merged = BucketBase::merge(
+                bm, cfg.LEDGER_PROTOCOL_VERSION, b1, b2,
+                /*shadows=*/{shadow},
+                /*keepTombstoneEntries=*/true,
+                /*countMergeEvents=*/true, clock.getIOContext(),
+                /*doFsync=*/true);
             EntryCounts e(merged);
             if (initEra)
             {
@@ -847,12 +848,12 @@ TEST_CASE_VERSIONS("merging bucket entries with initentry with shadows",
             // risking shadowing-out level 3. Level 4 is a placeholder here,
             // just to be a thing-to-merge-level-3-with in the presence of
             // shadowing from 1 and 2.
-            auto merge43 =
-                Bucket::merge(bm, cfg.LEDGER_PROTOCOL_VERSION, level4, level3,
-                              /*shadows=*/{level2, level1},
-                              /*keepTombstoneEntries=*/true,
-                              /*countMergeEvents=*/true, clock.getIOContext(),
-                              /*doFsync=*/true);
+            auto merge43 = BucketBase::merge(
+                bm, cfg.LEDGER_PROTOCOL_VERSION, level4, level3,
+                /*shadows=*/{level2, level1},
+                /*keepTombstoneEntries=*/true,
+                /*countMergeEvents=*/true, clock.getIOContext(),
+                /*doFsync=*/true);
             EntryCounts e43(merge43);
             if (initEra)
             {
@@ -873,12 +874,12 @@ TEST_CASE_VERSIONS("merging bucket entries with initentry with shadows",
 
             // Do a merge between level 2 and 1, producing potentially
             // an annihilation of their INIT and DEAD pair.
-            auto merge21 =
-                Bucket::merge(bm, cfg.LEDGER_PROTOCOL_VERSION, level2, level1,
-                              /*shadows=*/{},
-                              /*keepTombstoneEntries=*/true,
-                              /*countMergeEvents=*/true, clock.getIOContext(),
-                              /*doFsync=*/true);
+            auto merge21 = BucketBase::merge(
+                bm, cfg.LEDGER_PROTOCOL_VERSION, level2, level1,
+                /*shadows=*/{},
+                /*keepTombstoneEntries=*/true,
+                /*countMergeEvents=*/true, clock.getIOContext(),
+                /*doFsync=*/true);
             EntryCounts e21(merge21);
             if (initEra)
             {
@@ -899,13 +900,13 @@ TEST_CASE_VERSIONS("merging bucket entries with initentry with shadows",
 
             // Do two more merges: one between the two merges we've
             // done so far, and then finally one with level 5.
-            auto merge4321 =
-                Bucket::merge(bm, cfg.LEDGER_PROTOCOL_VERSION, merge43, merge21,
-                              /*shadows=*/{},
-                              /*keepTombstoneEntries=*/true,
-                              /*countMergeEvents=*/true, clock.getIOContext(),
-                              /*doFsync=*/true);
-            auto merge54321 = Bucket::merge(
+            auto merge4321 = BucketBase::merge(
+                bm, cfg.LEDGER_PROTOCOL_VERSION, merge43, merge21,
+                /*shadows=*/{},
+                /*keepTombstoneEntries=*/true,
+                /*countMergeEvents=*/true, clock.getIOContext(),
+                /*doFsync=*/true);
+            auto merge54321 = BucketBase::merge(
                 bm, cfg.LEDGER_PROTOCOL_VERSION, level5, merge4321,
                 /*shadows=*/{},
                 /*keepTombstoneEntries=*/true,
@@ -956,12 +957,12 @@ TEST_CASE_VERSIONS("merging bucket entries with initentry with shadows",
             // shadowing-out the init on level 3. Level 2 is a placeholder here,
             // just to be a thing-to-merge-level-3-with in the presence of
             // shadowing from 1.
-            auto merge32 =
-                Bucket::merge(bm, cfg.LEDGER_PROTOCOL_VERSION, level3, level2,
-                              /*shadows=*/{level1},
-                              /*keepTombstoneEntries=*/true,
-                              /*countMergeEvents=*/true, clock.getIOContext(),
-                              /*doFsync=*/true);
+            auto merge32 = BucketBase::merge(
+                bm, cfg.LEDGER_PROTOCOL_VERSION, level3, level2,
+                /*shadows=*/{level1},
+                /*keepTombstoneEntries=*/true,
+                /*countMergeEvents=*/true, clock.getIOContext(),
+                /*doFsync=*/true);
             EntryCounts e32(merge32);
             if (initEra)
             {
@@ -983,12 +984,12 @@ TEST_CASE_VERSIONS("merging bucket entries with initentry with shadows",
             // Now do a merge between that 3+2 merge and level 1, and we risk
             // collecting tombstones in the lower levels, which we're expressly
             // trying to _stop_ doing by adding INIT.
-            auto merge321 =
-                Bucket::merge(bm, cfg.LEDGER_PROTOCOL_VERSION, merge32, level1,
-                              /*shadows=*/{},
-                              /*keepTombstoneEntries=*/true,
-                              /*countMergeEvents=*/true, clock.getIOContext(),
-                              /*doFsync=*/true);
+            auto merge321 = BucketBase::merge(
+                bm, cfg.LEDGER_PROTOCOL_VERSION, merge32, level1,
+                /*shadows=*/{},
+                /*keepTombstoneEntries=*/true,
+                /*countMergeEvents=*/true, clock.getIOContext(),
+                /*doFsync=*/true);
             EntryCounts e321(merge321);
             if (initEra)
             {
