+++ src/history/FileTransferInfo.h
@@ -4,11 +4,10 @@
 // under the Apache License, Version 2.0. See the COPYING file at the root
 // of this distribution or at http://www.apache.org/licenses/LICENSE-2.0
 
-#include "bucket/Bucket.h"
+#include "bucket/LiveBucket.h"
 #include "crypto/Hex.h"
 #include "main/Config.h"
 #include "util/Fs.h"
-#include "util/Logging.h"
 #include "util/TmpDir.h"
 #include <string>
